Update POSsystem.h
percentage parameter data type change from int to float

+++ POSsystem.h
@@ -2,8 +2,8 @@
 #ifndef  POSSYSTEM_H
 #define POSSYSTEM_H
 
-  float calculateTip(float gross, int percentage);
-  float calculateNet(float gross, int percentage);
+  float calculateTip(float gross, float percentage);
+  float calculateNet(float gross, float percentage);
   void displayReciept(float gross, float net, float tip);
 
 #endif
